Improvements in and new features for MCMC optimiser.

+++ src/DiscreteModel.cpp
@@ -14,7 +14,7 @@ void NonLinearSRegDiscreteModel::set_parameters(myparam& PAR) {
     it=PAR.find("TriLikelihood"); m_triclique = boost::get<bool>(it->second);
     it=PAR.find("rescalelabels"); m_rescalelabels = boost::get<bool>(it->second);
     it=PAR.find("numthreads"); _nthreads = boost::get<int>(it->second);
-    it=PAR.find("labeldist"); _labeldist = boost::get<double>(it->second);
+    it=PAR.find("labeldist"); _labeldist = boost::get<float>(it->second);
     if(m_regoption == 1) _pairwise = true;
 }
